springboot-crud-rest-api[important]: Spring data JPA
In the last commit, we completed all the functionalities included in CRUD. And to make our system more flexible and robust, we introduce Service interface in our project structure. Which as bonus also provides us with an additional layer of abstraction.

Problem:
There is a problem in this approach. Or more like a room for improvement. So while integration of Service Layer, you may have noticed a lot of repititive work in between Dao and Service class. Which as a result increases the boiler plate code. For now we were just using one entity, but just consider a scenario, where there are some other entities like Salary, Department, etc. So to implement the basic CRUD operations for each entity, its just too much repititive work.

Solution:
To overcome the above mentioned problem, Springboot kinda provides us with magic. And this magic called Spring data JPA.

Detail:
You may have noticed, I am adding a new file in this commit i.e. EmployeeRepository and have commented both EmployeeDao and EmployeeDaoImpl class. Reason for commenting these classes is that we don't need them anymore. Instead EmployeeRepository class is doing all the heavy lifting for us. In this class I have extended this interface with JpaRespository. And it is taking two template parameters that are EntityType and PrimaryKey Type. And in return it is providing us with all the basic crud functionalities. Hence it reduce tremendous amount of boiler plate code.

We can provide the custom definition too for some database query related jobs. So it means we do not need Dao classes anymore.